add reset capability to contactinfo validator
refs #30446

there might be support cases where an administrator would like to
reset all in-memory data of the validator (and fill it with new data
afterwards). The reset part is done in this commit. The provisioning
of new data must be done by the application that consumes contargo-types
and its validators.