fix property mapping
refs #25562

it turns out that trying to set the property name in the constrain
validator context manually is futile. The how should you inside the
validation on which property the annoation was set.

the framework is handling this on its own (I guess via reflection or
something similar) no need to name the property at all. the code that
defined whether the property was 'mobile' or 'phone' did not work since
you cannot determie from an invalid number if it was a mobile or a (fixed)
phone number. and what about the fax numbers?

works now.