Fix: hibernate-validator dependency leaking into final artefact
The hibernate-validator dependency should not be part of the final
artefact, since it is only used to exercise the actual implementation of
the validators in the test cases. When using `contargo-types` with
modern projects, that might include the 6+ version of
hibernate-validator, it leads to actually two versions of the validator
on the final classpath. This is due to Maven groupId change in the
validator package: `org.hibernate:hibernate-validator` became
`org.hibernate.validator:hibernate-validator`.